ToDo LIst widget refactor
