+++ cogboard-webapp/src/components/widgets/types/ToDoListWidget/styled.js
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled/macro';
+
+import { Checkbox } from '@material-ui/core';
+
+export const StyledCheckbox = styled(Checkbox)`
+  margin-left: 15px;
+
+  &.Mui-checked {
+    color: rgba(255, 255, 255, 0.7);
+  }
+`;
